requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
emoji==2.10.1
Flask==3.0.3
fonttools==4.54.1
fpdf2==2.7.9
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.24.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.14.2
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
oauthlib==3.2.2
packaging==24.0
pillow==10.4.0
proto-plus==1.26.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.4
